Pass message to the worker without transferring buffer. Fixes #25
Reference:
https://github.com/kbumsik/opus-media-recorder/issues/25#issuecomment-636238714
Issue: https://github.com/kbumsik/opus-media-recorder/issues/25

The library is broken in Chrome 83. This is caused by the recent
update in the V8 engine: https://v8.dev/blog/v8-release-83

We stop transferring the buffer to mitigate the issue. However,
this leads to a dreased performance. We should figure out how to
transfer the buffers in a correct way.